Expose test cursor container internally
For testing purposes.
